Moved CrossSectionData out of the CrossSectionBlock
