+++ python/src/block/CrossSectionBlock.python.cpp
@@ -16,7 +16,7 @@ void wrapCrossSectionBlock( python::module& module, python::module& ) {
 
   // type aliases
   using Block = njoy::ACEtk::block::CrossSectionBlock;
-  using CrossSectionData = njoy::ACEtk::block::CrossSectionBlock::CrossSectionData;
+  using CrossSectionData = njoy::ACEtk::block::CrossSectionData;
 
   // wrap views created by this block
